Refactored namegiving of the utilities folder and classes.
