+++ EyeCT4Rails/EyeCT4RailsLogic/EyeCT4RailsLogic.csproj
@@ -52,11 +52,12 @@
     <Compile Include="Exceptions\InvalidUserException.cs" />
     <Compile Include="Exceptions\NoFreeSectionException.cs" />
     <Compile Include="Exceptions\UnknownException.cs" />
-    <Compile Include="Helpers\Hashing.cs" />
+    <Compile Include="Utilities\HashingUtil.cs" />
     <Compile Include="Repositories\MaintenanceRepository.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Repositories\RideManagementRepository.cs" />
     <Compile Include="Repositories\UserRepository.cs" />
+    <Compile Include="Utilities\SectionUtil.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\EyeCT4RailsDatabase\EyeCT4RailsDatabase.csproj">
