+++ EyeCT4Rails/EyeCT4RailsUnitTests/LogicTest/HashingTests.cs
@@ -1,4 +1,4 @@
-﻿using EyeCT4RailsLogic.Helpers;
+﻿using EyeCT4RailsLogic.Utilities;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace EyeCT4RailsUnitTests.LogicTest
@@ -12,8 +12,8 @@ namespace EyeCT4RailsUnitTests.LogicTest
             var salt = "G5F3SXd+Xnr8E3ncquQfzrQfhYgQ3UwXin59JTXJBc3BeJWM1bqIl2116v8W0cwznHMLb60mg/nAfWfbZJdzOA==";
             var password = "gekkehenkie";
 
-            var hash1 = Hashing.HashString(password, salt);
-            var hash2 = Hashing.HashString(password, salt);
+            var hash1 = HashingUtil.HashString(password, salt);
+            var hash2 = HashingUtil.HashString(password, salt);
 
             Assert.AreEqual(hash1,hash2);
         }
